Make FlowFile async-first and use a couple of Jasmine expectAsync()
