--- src/AsyncFlowFile.js
@@ -1,47 +1,0 @@
-import FlowFile from './FlowFile';
-
-/**
- * AsyncFlowFile class
- * @name AsyncFlowFile
- */
-export default class AsyncFlowFile extends FlowFile {
-
-  /**
-   * Retry aborted file upload
-   * @function
-   */
-  async retry() {
-    await this.bootstrap('retry');
-    return await this.flowObj.upload();
-  }
-
-  async bootstrap(event = null, initFileFn = this.flowObj.opts.initFileFn) {
-    /**
-     * Asynchronous initialization function, if defined, is run
-     * Then _bootstrap follow-up occurs
-     * And, optionally (in case of initial FlowFile creation), the `file-added` event is fired.
-     */
-    if (typeof initFileFn === 'function') {
-      await initFileFn(this, event);
-    }
-
-    this._bootstrap();
-
-    // console.log("Flowfile returns [async]", this._bootstrapped);
-    return this;
-  }
-
-  /**
-   * Indicates if string is being read at the moment
-   * @function
-   * @returns {boolean}
-   */
-  isReading() {
-    for (let chunk of this.chunks) {
-      if (chunk.status() === 'reading') {
-        return true;
-      }
-    }
-    return false;
-  }
-}
